Pick random words and check the answer
This commit randomly choose a word from the word_list and assigns it
to a variable called chosen_word.
It then asks the user to guess a letter and assign their answer to a
variable called guess.
Finally, it check if the letter the guess is one of the letters in
the chosen_word.